libs/puda-python/pyproject.toml
[project]
name = "puda-python"
version = "0.0.15"
description = "Communication library for the PUDA platform."
readme = "README.md"
authors = [
    { name = "zhao", email = "20024592+agentzhao@users.noreply.github.com" }
]
requires-python = ">=3.10"
dependencies = [
    "nats-py>=2.12.0",
    "pydantic>=2.12.5",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.uv.build-backend]
module-name = "puda"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]
